⚡️ Modularize typeDefs and resolvers
