+++ dashboard/server/src/lib/utils.ts
@@ -1,0 +1,7 @@
+import { NextFunction, Request, RequestHandler, Response } from 'express'
+
+export const catchErrors = (fn: RequestHandler) => (
+  req: Request,
+  res: Response,
+  next: NextFunction
+) => fn(req, res, next).catch(next)
