+++ dashboard/server/src/modules/Admin/index.ts
@@ -1,0 +1,4 @@
+import { typeDefs as adminTypeDefs } from './typeDefs'
+import { assignAdmin, createAdmin } from './resolvers'
+
+export { adminTypeDefs, assignAdmin, createAdmin }
