+++ dashboard/server/src/modules/Admin/typeDefs.ts
@@ -1,0 +1,15 @@
+const typeDefs = `
+  type Admin {
+    id: Int!
+    name: String!
+    image_url: String
+    tickets: [Ticket]
+  }
+
+  type Mutation {
+    createAdmin(name: String!, image_url: String): Admin
+    assignAdmin(adminId: Int!, ticketId: Int!): String
+  }
+`
+
+export { typeDefs }
