+++ dashboard/server/src/modules/Space/index.ts
@@ -1,0 +1,4 @@
+import { typeDefs as spaceTypeDefs } from './typeDefs'
+import { spaces } from './resolvers'
+
+export { spaces, spaceTypeDefs }
