+++ dashboard/server/src/modules/Space/typeDefs.ts
@@ -1,0 +1,12 @@
+const typeDefs = `
+  type Space {
+    id: Int!
+    name: String
+  }
+
+  type Query {
+    spaces: [Space]
+  }
+`
+
+export { typeDefs }
