+++ dashboard/server/src/modules/Ticket/index.ts
@@ -1,0 +1,4 @@
+import { typeDefs as ticketTypeDefs } from './typeDefs'
+import { tickets } from './resolvers'
+
+export { tickets, ticketTypeDefs }
