+++ dashboard/server/src/modules/Ticket/typeDefs.ts
@@ -1,0 +1,26 @@
+const typeDefs = `
+  type Ticket {
+    id: Int!
+    title: String!
+    body: String!
+    createdAt: String
+    assignedTo: Admin
+    author: User
+    topics: [Topic]
+  }
+
+  type TicketsResponse {
+    tickets: [Ticket]
+    totalCount: Int
+  }
+
+  extend type Query {
+    tickets(spaceId: Int): TicketsResponse
+  }
+
+  schema {
+    query: Query
+  }
+`
+
+export { typeDefs }
