+++ dashboard/server/src/modules/Topic/index.ts
@@ -1,0 +1,3 @@
+import { typeDefs as topicTypeDefs } from './typeDefs'
+
+export { topicTypeDefs }
