+++ dashboard/server/src/modules/Topic/typeDefs.ts
@@ -1,0 +1,8 @@
+const typeDefs = `
+  type Topic {
+    id: Int
+    name: String
+  }
+`
+
+export { typeDefs }
