+++ dashboard/server/src/modules/User/index.ts
@@ -1,0 +1,3 @@
+import { typeDefs as userTypeDefs } from './typeDefs'
+
+export { userTypeDefs }
