+++ dashboard/server/src/modules/User/typeDefs.ts
@@ -1,0 +1,9 @@
+const typeDefs = `
+  type User {
+    id: Int
+    username: String
+    reputation: Int
+  }
+`
+
+export { typeDefs }
